\name{conexus-package}

\alias{conexus-package}

\alias{conexus}

\docType{package}

\title{Evaluation of ex-situ conservation status of species}

\description{
This package implements a method to evaluate the ex-situ conservation status of a species. The method is described in: 

El Graoui, M., M.E. Ghanem, M. Amri, R.J. Hijmans, 2025. A distance-based framework for assessing the ex-situ conservation status of plants. Submitted for publication. 
}

\section{Authors}{
Marwa el Graoui and Robert J. Hijmans
}


\keyword{ package }

